add zod schemas to core
