+++ packages/frame-core/src/index.ts
@@ -1,1 +1,2 @@
 export * from "./types";
+export * from "./schemas";
