+++ packages/frame-core/src/schemas/embeds.ts
@@ -1,0 +1,29 @@
+import { z } from 'zod';
+import { colorValueSchema, frameNameSchema, secureUrlSchema } from './shared';
+
+export const actionLaunchFrameSchema = z.object({
+  type: z.literal('launch_frame'),
+  name: frameNameSchema,
+  url: secureUrlSchema,
+  splashImageUrl: secureUrlSchema.optional(),
+  splashBackgroundColor: colorValueSchema.optional(),
+});
+
+export const actionSchema = z.discriminatedUnion('type', [actionLaunchFrameSchema]);
+
+export const buttonTitleSchema = z.string().max(32);
+
+export const buttonSchema = z.object({
+  title: buttonTitleSchema,
+  action: actionSchema,
+});
+
+export const frameEmbedNextSchema = z.object({
+  version: z.literal('next'),
+  imageUrl: secureUrlSchema,
+  button: buttonSchema,
+});
+
+export const safeParseFrameEmbed = (rawResponse: unknown) => frameEmbedNextSchema.safeParse(rawResponse);
+
+export type FrameEmbedNext = z.infer<typeof frameEmbedNextSchema>;
